requirements.txt
netaddr==0.8.0
netifaces==0.10.9
pyalsaaudio==0.9.0
